# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AdditionalLoad:
    properties:
      comment:
        description: Комментарий
        type: string
      load_id:
        description: ID дополнительной нагрузки
        format: uuid
        type: string
      name:
        description: Название нагрузки
        type: string
      volume:
        description: Объем
        type: string
    type: object
...
  models.ClassroomLoad:
    properties:
      group_name:
        description: Название группы
        type: string
      hours:
        description: Кол-во часов
        type: integer
      load_id:
        description: ID аудиторной нагрузки
        format: uuid
        type: string
      load_type:
        description: Тип аудиторной нагрузки
        enum:
        - practice
        - lectures
        - laboratory
        type: string
      main_teacher:
        description: Основное учитель
        type: string
      subject_name:
        description: Название предмета
        type: string
    type: object
  models.Conference:
    properties:
      conference_id:
        description: ID конференции
        format: uuid
        type: string
      conference_name:
        description: Название конференции
        type: string
      location:
        description: Место проведения
        type: string
      report_name:
        description: Название доклада
        type: string
      reported_at:
        description: Дата доклада
        format: date-time
        type: string
      rinc:
        description: Индекс РИНЦ
        type: boolean
      scopus:
        description: Индекс Scopus
        type: boolean
      status:
        description: Статус прогресса научной конференции
        enum:
        - registered
        - performed
        type: string
      wac:
        description: Индекс WaC
        type: boolean
      wos:
        description: Индекс WoS
        type: boolean
    type: object
...
  models.Group:
    properties:
      group_id:
        type: integer
      name:
        type: string
    type: object
  models.IndividualStudentsLoad:
    properties:
      comment:
        description: Комментарий
        type: string
      load_id:
        description: ID индивидуальной работы со студентами
        format: uuid
        type: string
      load_type:
        description: Тип индивидуальной работы
        enum:
        - project practice
        - bachelor
        - masters
        type: string
      students_amount:
        description: Количество студентов
        type: integer
    type: object
...
  models.Specialization:
    properties:
      name:
        type: string
      specialization_id:
        type: integer
    type: object
...
  request_models.AddGroupsRequest:
    properties:
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
    type: object
  request_models.AddSpecializationsRequest:
    properties:
      specs:
        items:
          $ref: '#/definitions/models.Specialization'
        type: array
    type: object
...
  request_models.DeleteEnumRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  request_models.DeleteIDs:
    properties:
      ids:
        items:
          type: string
        type: array
      semester:
        type: integer
    type: object
...
  request_models.UpsertAdditionalLoadRequest:
    properties:
      loads:
        description: |-
          Дополнительные нагрузки, которые нужно добавить/изменить. Если запрос на добавление, то
          ID дополнительной нагрузки (load_id) не заполнять (делать null), в противном случае - иначе
        items:
          $ref: '#/definitions/models.AdditionalLoad'
        type: array
      semester:
        description: Семестр
        type: integer
      t_load_id:
        description: ID совокупности всех нагрузок за семестр
        format: uuid
        type: string
    type: object
  request_models.UpsertClassroomLoadRequest:
    properties:
      loads:
        description: |-
          Аудиторные нагрузки, которые нужно добавить/изменить. Если запрос на добавление, то
          ID нагрузки (load_id) не заполнять (делать null), в противном случае - иначе
        items:
          $ref: '#/definitions/models.ClassroomLoad'
        type: array
      semester:
        description: Семестр
        type: integer
      t_load_id:
        description: ID совокупности всех нагрузок за семестр
        format: uuid
        type: string
    type: object
  request_models.UpsertConferencesRequest:
    properties:
      conferences:
        description: |-
          Научные конференции, которые нужно добавить/изменить. Если запрос на добавление, то
          ID конференции (conference_id) не заполнять (делать null), в противном случае - иначе
        items:
          $ref: '#/definitions/models.Conference'
        type: array
      semester:
        description: Семестр
        type: integer
      work_id:
        description: ID совокупности всех научных работ за семестр
        format: uuid
        type: string
    type: object
...
  request_models.UpsertIndividualLoadRequest:
    properties:
      loads:
        description: |-
          Индивидуальные нагрузки, которые нужно добавить/изменить. Если запрос на добавление, то
          ID индивидуальной нагрузки (load_id) не заполнять (делать null), в противном случае - иначе
        items:
          $ref: '#/definitions/models.IndividualStudentsLoad'
        type: array
      semester:
        type: integer
      t_load_id:
        format: uuid
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Сервис Kasper
  version: 0.1.2
paths:
  /administrator/enum/groups/{token}:
    get:
      consumes:
      - application/json
      description: Получение списка всех групп
      parameters:
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Получение списка всех групп
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Добавление групп
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.AddGroupsRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Добавление групп
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Удаление групп
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.DeleteEnumRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Удаление групп
      tags:
      - Admin
  /administrator/enum/specializations/{token}:
    get:
      consumes:
      - application/json
      description: Получение списка всех групп
      parameters:
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные
          schema:
            items:
              $ref: '#/definitions/models.Specialization'
            type: array
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Получение списка всех групп
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Добавление групп
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.AddSpecializationsRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Добавление групп
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Удаление специализаций
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.DeleteEnumRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Удаление специализаций
      tags:
      - Admin
...
  /students/load/additional/{token}:
    post:
      consumes:
      - application/json
      description: Добавление или обновление дополнительной нагрузки
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.UpsertAdditionalLoadRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Добавление или обновление дополнительной нагрузки
      tags:
      - Student.TeachingLoad
    put:
      consumes:
      - application/json
      description: Удаление дополнительной нагрузки по ID нагрузки
      parameters:
      - description: ID нагрузок и семестр
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.DeleteIDs'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Удаление дополнительной педнагрузки
      tags:
      - Student.TeachingLoad
  /students/load/classroom/{token}:
    post:
      consumes:
      - application/json
      description: Добавление или обновление аудиторной нагрузки
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.UpsertClassroomLoadRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Добавление или обновление аудиторной нагрузки
      tags:
      - Student.TeachingLoad
    put:
      consumes:
      - application/json
      description: Удаление аудиторной нагрузки по ID нагрузки
      parameters:
      - description: ID нагрузок и семестр
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.DeleteIDs'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Удаление аудиторной педнагрузки
      tags:
      - Student.TeachingLoad
  /students/load/individual/{token}:
    post:
      consumes:
      - application/json
      description: Добавление или обновление индивидуальных нагрузок
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.UpsertIndividualLoadRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Добавление или обновление индивидуальных нагрузок
      tags:
      - Student.TeachingLoad
    put:
      consumes:
      - application/json
      description: Удаление индивидуальных педнагрузок по ID нагрузок
      parameters:
      - description: ID нагрузок и семестр
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.DeleteIDs'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Удаление индивидуальных педагогических нагрузок
      tags:
      - Student.TeachingLoad
...
  /students/works/conferences/{token}:
    post:
      consumes:
      - application/json
      description: Добавление или обновление научных конференций
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.UpsertConferencesRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Добавление или обновление научных конференций
      tags:
      - Student.ScientificWorks
    put:
      consumes:
      - application/json
      description: Удаление научных конференций по ID конференций
      parameters:
      - description: ID работ и семестр
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.DeleteIDs'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Удаление научных конференций
      tags:
      - Student.ScientificWorks
...
swagger: "2.0"